Check to see if we could actually find a release

+++ PFHelpers/Install-PSGithubRelease.ps1
@@ -13,28 +13,34 @@ Param(
 
     if ($PreRelease) {
         $releasesUri = "https://api.github.com/repos/$GithubRepo/releases"
-        $downloadUri = ((Invoke-RestMethod -Method GET -Uri $releasesUri)[0].assets | Where-Object name -like $filenamePattern ).browser_download_url
+        $downloadUri = ((Invoke-RestMethod -ErrorAction SilentlyContinue -Method GET -Uri $releasesUri)[0].assets | Where-Object name -like $filenamePattern ).browser_download_url
     }
     else {
         $releasesUri = "https://api.github.com/repos/$GithubRepo/releases/latest"
-        $downloadUri = ((Invoke-RestMethod -Method GET -Uri $releasesUri).assets | Where-Object name -like $filenamePattern ).browser_download_url
+        $downloadUri = ((Invoke-RestMethod -ErrorAction SilentlyContinue -Method GET -Uri $releasesUri).assets | Where-Object name -like $filenamePattern ).browser_download_url
     }
 
-    $pathZip = Join-Path -Path $([System.IO.Path]::GetTempPath()) -ChildPath $(Split-Path -Path $downloadUri -Leaf)
+    if($downloadUri)
+    {
+        $pathZip = Join-Path -Path $([System.IO.Path]::GetTempPath()) -ChildPath $(Split-Path -Path $downloadUri -Leaf)
 
-    Invoke-WebRequest -Uri $downloadUri -Out $pathZip
+        Invoke-WebRequest -Uri $downloadUri -Out $pathZip
 
-    Remove-Item -Path $pathExtract -Recurse -Force -ErrorAction SilentlyContinue
+        Remove-Item -Path $pathExtract -Recurse -Force -ErrorAction SilentlyContinue
 
-    if ($innerDirectory) {
-        $tempExtract = Join-Path -Path $([System.IO.Path]::GetTempPath()) -ChildPath $((New-Guid).Guid)
-        Expand-Archive -Path $pathZip -DestinationPath $tempExtract -Force
-        Move-Item -Path "$tempExtract\*" -Destination $pathExtract -Force
-        Remove-Item -Path $tempExtract -Force -Recurse -ErrorAction SilentlyContinue
+        if ($innerDirectory) {
+            $tempExtract = Join-Path -Path $([System.IO.Path]::GetTempPath()) -ChildPath $((New-Guid).Guid)
+            Expand-Archive -Path $pathZip -DestinationPath $tempExtract -Force
+            Move-Item -Path "$tempExtract\*" -Destination $pathExtract -Force
+            Remove-Item -Path $tempExtract -Force -Recurse -ErrorAction SilentlyContinue
+        }
+        else {
+            Expand-Archive -Path $pathZip -DestinationPath $pathExtract -Force
+        }
+
+        Remove-Item $pathZip -Force
     }
     else {
-        Expand-Archive -Path $pathZip -DestinationPath $pathExtract -Force
+        Write-Warning "Could not get latest release, please check the repo has an available release and try again ($GithubRepo)"
     }
-
-    Remove-Item $pathZip -Force
